SAK-10244 Removing hard coded references to the news tool and web tool from the POH and replacing with a sakai.properties setting that takes in a coma delimted list of tool ids (just like the stealth tools list does) and allows configuring at the system level which tools should be allowed to be added multiple times to the same site. So for instance adding the following line to your sakai.properties file would allow news, web and the spiffy contrib content-viewer tool to be added multiple times per site.
sakai.site.multiPlacementTools=sakai.news,sakai.iframe,sakai.contentviewer

Leaving this setting un set will just use the default of web and news tools like it did before.

In the future it would be nice to define the defaults somewhere else, such as a static string in a tool api or something.


git-svn-id: https://source.sakaiproject.org/svn/site-manage/trunk@31280 66ffb92e-73f9-0310-93c1-f5514f145a0a